+++ package.json
@@ -37,6 +37,7 @@
     "react-dom": "19.1.0",
     "react-error-boundary": "^5.0.0",
     "react-hook-form": "^7.55.0",
+    "react-use-wizard": "^2.3.0",
     "server-only": "^0.0.1",
     "zod": "^3.24.2",
     "zod-i18n-map": "^2.27.0"
