--- src/app/(app)/verify/guilds/[guildId]/form-schema.ts
@@ -1,5 +1,0 @@
-﻿import { z } from 'zod';
-
-export const captchaFormSchema = z.object({
-  turnstileToken: z.string(),
-});
