+++ src/app/(app)/verify/guilds/[guildId]/loading.tsx
@@ -5,7 +5,7 @@ import { Spinner } from '@heroui/spinner';
 
 export default function Loading() {
   return (
-    <Card className='p-8'>
+    <Card className='p-8 w-fit'>
       <Spinner variant='spinner' />
     </Card>
   );
