+++ src/app/(app)/verify/guilds/[guildId]/page.tsx
@@ -1,7 +1,10 @@
-﻿import { getGuild, isUserJoinedGuild } from '@/lib/discord/api';
+﻿import { Logo } from '@/components/logo';
+import { auth } from '@/lib/auth';
+import { getGuild, getGuildMember, isUserJoinedGuild } from '@/lib/discord/api';
 import { db } from '@/lib/drizzle';
+import { Card } from '@heroui/card';
 import { forbidden, notFound } from 'next/navigation';
-import MultiStepVerificationCard from './multistep-form-card';
+import { VerificationWizard } from './wizard';
 
 export default async function Page({
   params,
@@ -15,12 +18,21 @@ export default async function Page({
   const guild = await getGuild(guildId).catch(() => null);
   if (!guild) notFound();
 
+  const session = await auth();
   const setting = await db.query.verificationSetting.findFirst({
     where: (setting, { eq }) => eq(setting.guildId, guild.id),
   });
 
   if (!setting?.enabled || !setting.role || setting.captchaType !== 'web') notFound();
-  if (!(await isUserJoinedGuild(guild.id))) forbidden();
+  if (!session || !(await isUserJoinedGuild(guild.id))) forbidden();
 
-  return <MultiStepVerificationCard guild={guild} />;
+  const member = await getGuildMember(guild.id, session.user.id);
+  const isVerified = member.roles.some((roleId) => roleId === setting.role);
+
+  return (
+    <Card className='w-full flex flex-col gap-6 px-6 py-8'>
+      <Logo width={110} />
+      <VerificationWizard guild={guild} isVerified={isVerified} />
+    </Card>
+  );
 }
