+++ src/app/(app)/verify/guilds/[guildId]/schema.ts
@@ -1,0 +1,8 @@
+﻿import { z } from 'zod';
+
+export type FormInputSchema = z.input<typeof captchaFormSchema>;
+export type FormOutputSchema = z.input<typeof captchaFormSchema>;
+
+export const captchaFormSchema = z.object({
+  turnstileToken: z.string(),
+});
